fix: test for playability without user
