+++ src/index.ts
@@ -79,6 +79,7 @@ const routes = [
   "trackGroups/{id}/wishlist",
   "trackGroups/{id}/emailDownload",
   "trackGroups/{id}/generate",
+  "playable",
   "tracks",
   "tracks/{id}",
   "tracks/{id}/audio",
@@ -112,7 +113,6 @@ const routes = [
   "users/{userId}/notifications/unreadCount",
   "users/{userId}/notifications/{notificationId}",
   "users/{userId}/trackGroupPurchases",
-  "users/{userId}/tracksPlayable",
   "users/{userId}/purchases",
   "users/{userId}/wishlist",
   "users/{userId}/charges",
